Pull datasets table into separate component
