+++ context/app/static/js/components/workspaces/AddDatasetsTable/index.ts
@@ -1,0 +1,3 @@
+import AddDatasetsTable from './AddDatasetsTable';
+
+export default AddDatasetsTable;
